(change) Saving user on base

+++ src/modules/base/service/base.service.ts
@@ -7,14 +7,14 @@ import {
 } from '@nestjs/common';
 import { LoggerServer } from 'src/loggerServer';
 
-import { BaseDTO } from 'src/DTO/base.dto';
+import { BaseDTO } from '../interfaces/base.dto';
 
-import * as csvConverter from 'csvjson-csv2json';
-import { Base } from 'src/Mongo/Interface/base.interface';
-import { FilterDTO } from 'src/Mongo/Interface/query.interface';
-import { BaseRepository } from 'src/Mongo/repository/base.repository';
+import { parse } from 'papaparse';
+import { FilterDTO } from 'src/interfaces/query.interface';
 import { ParametersService } from 'src/modules/parameters/services/parameters.service';
 import { StructureService } from 'src/modules/structure/service/structure.service';
+import { BaseRepository } from '../base.repository';
+import { Base } from '../interfaces/base.interface';
 
 @Injectable()
 export class BaseService {
@@ -33,24 +33,25 @@ export class BaseService {
         `Arquivo grande, pode demorar mais que o normal. Tamanho: ${file.size} bytes aproximadamente | Nome: ${file.originalname}`,
       );
 
-    return csvConverter(file.buffer.toString('utf-8'), {
-      parseNumbers: true,
-    });
+    return parse(file.buffer.toString('utf-8'), {
+      delimiter: ';',
+      dynamicTyping: true,
+      header: true,
+    }).data;
   }
 
   async uploadFiles(
     files: Array<Express.Multer.File>,
-    structure: string,
+    structureID: string,
     name: string,
+    userID: string,
   ) {
     if (files === undefined || files.length === 0) {
       this.logger.error('Enviado arquivos inválidos ou nenhum arquivo enviado');
       throw new HttpException('No files uploaded', HttpStatus.BAD_REQUEST);
     }
 
-    structure = structure.toLowerCase();
-
-    const structureObject = this.structureService.getStructureByName(structure);
+    const structureObject = await this.structureService.getByID(structureID);
     if (!structureObject)
       throw new HttpException(
         'Estrutura não encontrada',
@@ -67,31 +68,46 @@ export class BaseService {
       `Fazendo o upload dos arquivos e convertendo para JSON. Número de arquivos: ${files.length}`,
     );
 
-    //Isso permite com que o sistema possa adicionar novos tipos de simulação
-    const structureFinal = new BaseDTO();
+    const structureFinal = {} as Base;
     structureFinal.name = name;
-    structureFinal.parameters = {} as any;
-    structureFinal.type = structure as any;
-    const params = Object.keys(structureObject.type_parameters);
+    structureFinal.type = structureID;
+    structureFinal.user = userID;
+    structureFinal.parameters = structureObject.parameters;
 
-    for (let i = 0; i < params.length; i++) {
-      const param = params[i];
-      const subParams = Object.keys(structureObject.type_parameters[param]);
+    const params = Object.keys(structureObject.parameters);
 
-      if (subParams.length === 0)
-        structureFinal.parameters[param] = this.convertJSON(files[i]);
-      else {
-        for (const subParam of subParams) {
-          structureFinal.parameters[param][subParam] = this.convertJSON(
-            files[i],
-          );
-          i++;
+    try {
+      for (let i = 0; i < params.length; i++) {
+        const param = params[i];
+        const subParams = Object.keys(structureObject.parameters[param]);
+
+        if (subParams.length === 0)
+          structureFinal.parameters[param] = this.convertJSON(files[i]);
+        else {
+          for (const subParam of subParams) {
+            structureFinal.parameters[param][subParam] = this.convertJSON(
+              files[i],
+            );
+
+            i++;
+          }
         }
       }
+    } catch (e) {
+      throw new HttpException(
+        'Erro ao converter os arquivos! Valide se os arquivos estão no formato correto',
+        HttpStatus.INTERNAL_SERVER_ERROR,
+      );
     }
 
     try {
-      return await this.baseRepository.saveBase(structureFinal);
+      const parameters = await this.parametersService.uploadAllParameters(
+        structureFinal.parameters,
+      );
+      return await this.baseRepository.saveBase({
+        ...structureFinal,
+        parameters,
+      });
     } catch (e) {
       throw new HttpException(
         `Erro ao salvar a base ${e}`,
@@ -100,7 +116,7 @@ export class BaseService {
     }
   }
 
-  async getBases(query: FilterDTO): Promise<Base[]> {
+  async getBases(query: FilterDTO) {
     return await this.baseRepository.getBases(query);
   }
 
@@ -116,25 +132,57 @@ export class BaseService {
     return base;
   }
 
-  async saveBase({ parameters, ...base }: BaseDTO): Promise<Base> {
-    const parametersObject =
-      await this.parametersService.uploadAllParameters(parameters);
+  async saveBase(
+    { parameters, ...base }: BaseDTO,
+    structureID: string,
+    userID: string,
+  ): Promise<Base> {
+    const structure = await this.structureService.getByID(structureID);
+    if (!structure)
+      throw new HttpException(
+        'Estrutura não encontrada',
+        HttpStatus.BAD_REQUEST,
+      );
 
-    return await this.baseRepository.saveBase({
-      ...base,
-      parameters: parametersObject,
-    });
+    try {
+      const parametersObject =
+        await this.parametersService.uploadAllParameters(parameters);
+
+      return await this.baseRepository.saveBase({
+        ...base,
+        user: userID,
+        type: structureID,
+        parameters: parametersObject,
+      });
+    } catch (e) {
+      this.logger.error(`Erro ao salvar a base! Erro: ${e}`);
+
+      throw new HttpException(
+        'Erro ao salvar base',
+        HttpStatus.INTERNAL_SERVER_ERROR,
+      );
+    }
   }
 
-  async updateBase(baseID: string, newBase: BaseDTO): Promise<Base> {
-    try {
-      const base = await this.baseRepository.getBaseByID(baseID);
-      if (!base)
-        throw new HttpException(
-          'Nenhuma base encontrada com esse ID',
-          HttpStatus.NOT_FOUND,
-        );
+  async updateBase(
+    baseID: string,
+    newBase: BaseDTO,
+    userID: string,
+  ): Promise<Base> {
+    const base = await this.baseRepository.getBaseByID(baseID);
+    if (!base)
+      throw new HttpException(
+        'Nenhuma base encontrada com esse ID',
+        HttpStatus.NOT_FOUND,
+      );
 
+    if (base.user !== userID)
+      throw new HttpException(
+        'Você não tem permissão para editar essa base',
+        HttpStatus.UNAUTHORIZED,
+      );
+
+    try {
       const parametersObject = await this.parametersService.uploadAllParameters(
         newBase.parameters,
       );
@@ -155,13 +203,29 @@ export class BaseService {
     }
   }
 
-  async deleteBase(baseID: string): Promise<Base> {
+  async deleteBase(baseID: string, userID: string): Promise<Base> {
+    const base = await this.baseRepository.getBaseByID(baseID);
+
+    if (!base)
+      throw new HttpException(
+        'Nenhuma base encontrada com esse ID',
+        HttpStatus.NOT_FOUND,
+      );
+
+    if (base.user.toString() !== userID)
+      throw new HttpException(
+        'Você não tem permissão para deletar essa base',
+        HttpStatus.UNAUTHORIZED,
+      );
+
     try {
       const baseDeleted = await this.baseRepository.deleteBase(baseID);
+      console.debug('baseDeleted', baseDeleted);
 
       this.logger.warn(`Base deletada: ${baseDeleted.name}`);
       return baseDeleted;
     } catch (e) {
+      this.logger.error(`Erro ao deletar a base! Erro: ${e}`);
       throw new HttpException(
         'Algum erro ocorreu ao deletar a base',
         HttpStatus.NOT_FOUND,
